+++ datasets.tsv
@@ -79,3 +79,4 @@ Martinez-2025-AffectiveRatings	Martínez, Gonzalo and Molero, Juan Diego and Gon
 Cameirao-2010-AoA	Cameirão, Manuela L. and Vicente, Selene G.	2010	ratings	Portuguese	Portuguese	https://doi.org/10.3758/BRM.42.2.474	Cameirao2010	This list contains age of acquisition ratings for 1,749 Portuguese words. Participants were asked to rate the age at which they thought they had learned a given word on a 9-point scale ranging from 2 years old to +13 years old. The dataset also includes ratings for familiarity ([Marques 2004](:bib:Marques2004)), concreteness and imageability ([Marques 2005](:bib:Marques2005)) which were added here. The words in the original dataset are capitalized, resulting in the format in the present list.
 SanMiguelAbella-2020-MotorContent	San Miguel Abella, Romina A. and González-Nosti, María	2020	ratings	Spanish	Spanish	https://doi.org/10.3758/s13428-019-01241-1	SanMiguelAbella2020	This list contains motor content ratings, i.e. the amount of mobility an actions entails, for 4,565 Spanish verbs given on a 7-point scale. The dataset further includes age of acquisition data from [Alonso et al. 2016](:bib:Alonso2016) which is already included in NoRaRe and not duplicated here. Additionally, the original dataset includes data on frequency, familiarity, imageability and concreteness, sourced from the EsPal database ([Duchon et al. 2013](:bib:Duchon2013)), which were not included here. The verbs in the original dataset are capitalized, resulting in the format in the present list.
 Grandy-2020-EmotionImageability	Grandy, Thomas H. and Lindenberger, Ulman and Schmiedek, Florian	2020	ratings	German	German	https://doi.org/10.3758/s13428-019-01294-2	Grandy2020	This list contains ratings of emotionality and imageability of about 2500 German nouns given by younger (21-31 years old) and older (70-86 years old) adults. Ratings for imageability were given on a 100-point sliding scale while ratings for emotionality were given on a 200-point sliding scale. Participants used their mouse cursor to select or slide to a point on the scale, the exact numeric value of their selection was displayed for reference.
+Winter-2022-SemanticChange	Winter, Bodo and Srinivasan, Mahesh	2022	relations	English	English	https://doi.org/10.1080/10926488.2021.1945419	Winter2022	This network contains asymmetrical, cross-linguistically attested semantic changes between semantically related concepts. It is based on the study by [Urban 2011](:bib:Urban2011).
