+++ datasets/Winter-2022-SemanticChange/Winter-2022-SemanticChange.tsv-metadata.json
@@ -1,69 +1,66 @@
 {
     "@context": [
-        "http://www.w3.org/ns/csvw", 
+        "http://www.w3.org/ns/csvw",
         {
             "@language": "en"
         }
-    ], 
+    ],
     "dialect": {
-        "delimiter": "\t", 
+        "encoding": "utf-8-sig",
+        "delimiter": "\t",
         "skipBlankRows": true
-    }, 
+    },
     "tables": [
         {
             "tableSchema": {
                 "columns": [
                     {
                         "datatype": {
-                            "base": "string", 
+                            "base": "string",
                             "format": "[a-zA-Z]+\\-[0-9]{4}\\-[0-9]+[a-z]?\\-[0-9]+[a-z]?$"
-                        }, 
+                        },
                         "name": "ID"
-                    }, 
+                    },
                     {
                         "datatype": {
-                            "base": "string", 
+                            "base": "string",
                             "format": "[0-9\\.]+([a-z\\\u2013]+)?$"
-                        }, 
+                        },
                         "name": "NUMBER"
-                    }, 
+                    },
                     {
                         "datatype": {
-                            "base": "integer", 
+                            "base": "integer",
                             "minimum": 1
-                        }, 
+                        },
                         "name": "CONCEPTICON_ID"
-                    }, 
+                    },
                     {
-                        "datatype": "string", 
+                        "datatype": "string",
                         "name": "CONCEPTICON_GLOSS"
-                    }, 
+                    },
                     {
-                        "datatype": "string", 
+                        "datatype": "string",
                         "name": "ENGLISH"
-                    }, 
+                    },
                     {
-                        "datatype": "string", 
-                        "name": "SEMANTIC_CLASS_ID"
-                    }, 
-                    {
-                        "datatype": "string", 
-                        "name": "CATEGORY"
-                    }, 
-                    {
-                        "datatype": "json", 
+                        "datatype": "json",
                         "name": "TARGET_CONCEPTS"
                     },
                     {
-                        "datatype": "json", 
+                        "datatype": "json",
+                        "name": "SOURCE_CONCEPTS"
+                    },
+                    {
+                        "datatype": "json",
                         "name": "LINKED_CONCEPTS"
                     }
-                ], 
+                ],
                 "primaryKey": [
                     "ID"
                 ]
-            }, 
-            "url": "Winter-2022-SemanticShift.tsv"
+            },
+            "url": "Winter-2022-SemanticChange.tsv"
         }
     ]
 }
